requirements/start.txt
pydantic==2.12.5
PyYAML==6.0.3
pandas==2.3.3